basic deoptimization infrastructure
This commit adds a new file named deoptimize.c, which basically saves
yet-to-optimize instruction sequence aside, and reverts to that state
on occasions.

The "state" here means the ISeq's raw sequence.  In other words the
"optimization" assumes that everything other than the sequence are
shared among optimized / non-optimized variant.  Examples of such info
includes for instance program counters and catch tables; so under this
infrastructure you cannot optimize an iseq which deals exceptions yet.
That is to be developed later.

There is no corresponding on-the-fly optimizer now; this commit only
adds overheads.  If you take a benchmark on this commit it should be
(hopefully slightly) slower than the original.  Any optimizations that
utilizes this mechanism must be sufficiently faster than the overhead
introduced in this commit.

	* deoptimize.h: new file.

	* deoptimize.c: new file.

	* common.mk (COMMONOBJS): dependencies for new files.

	* vm_core.h (rb_vm_struct): new field.

	* vm_core.h (rb_iseq_constant_body): new field.

	* iseq.c (rb_iseq_free): proper handling of new field.

	* iseq.c (iseq_memsize): ditto.