basic deoptimization infrastructure.
This commit adds a new file named deoptimize.c, which basically saves
yet-to-optimize instruction sequence aside, and reverts to that state on
occasions.

The "state" here means the set of ISeq raw sequence and program counter.
Or, the "optimization" assumes that everything other than that two are
shared among optimized / non-optimized variant.  Examples of such info
includes for instance catch table; so under this infrastructure you
cannot optimize an iseq which deals exceptions yet.  That is to be
developed later.

There is no corresponding on-the-fly optimizer now; this commit only
adds overheads.  If you take a benchmark on this commit it should be
(hopefully slightly) slower than the original.  Any optimizations that
utilizes this mechanism must be sufficiently faster than the overhead
introduced in this commit.