(FAILED) avoid function call
All calls to rb_intern() results in this function.  All those calls result in
rb_usascii_encoding() function call.

But wait.  A rb_encoding* variable is designed to be nullable, and a NULL encoding
means ASCII encoding.  Is there any reason to call that function instead?

It is in fact, no.  Passing NULL here perfectly works as it ought.  The problem is,
it boosts nothing.  This overhead to call extern function is very small that I
cannot measure any difference in any benchmark.