define purity of each instructions
An instruction is said to be pure if it concerns nothing but its
operands and stack.  This concept is much like GCC's
__attribute__((__pure__)).  To put it another way, instructions that
share any of following characteristics are _not_ pure:

- Either reads or writes a global, class, or instance variable.
- Defines a constant.
- Interacts with other threads.
- Calls rb_funcall inside (can't say purity for C).
- Evaluates other ISeqs that include non-pure instructions.

The last one is complicated.  Typically you cannot say if a method is
written in either Ruby or C, before you actually call that method.  A
good news is however, that once a method is called, last-called method
definition is cached inside of an instruction operand. A method's
purity can be detectable in that case.

So this purity is not something static, nor simple "pure / not pure"
distinction.  In fact an instruction's purity vary from time to time
on the fly, and in order to do so we define following four kinds of
purity:

- The instruction is definitely pure.
- The instruction is definitely not pure.
- The instruction's purity is not fixed yet; maybe because that
  instruction has never been evaluated before.
- The instruction's purity depends on a block parameter.

For the last kind, consider this ruby script:

    def foo
      return yield self
    end

Here, this method foo can either be pure or not pure, depending on how
the yield works, but nothing other than that.  This kind of situation
is something we want to optimize.  We are distinguishing this from
simply being not pure.