rb_id_table must come with mark function.
The struct rb_id_table can hold arbitrary VALUE values.  Now that this
struct is reachable form Ruby's object space, it must understand what
GC requests to it.

Note that the implementation is in gc.c. I heared that this is better
placed here than in id_table.c, because this way we can minimize
access to rb_objspace global bariable.