hotspot detection & update purity using it
An instruction sequence's purity varys time to time.  Updating it
needs recursive scans of the sequence's children.  This is not always
lightweight.  In order to limit unnecessary scans over and over again,
we are introducing hotspot detection.

A hotspot is detected per method call site.  Every time a valid call
cache is hit, it increments its temperature.  Once it reached max it
is considered "hot enough", then the calling method's purity is
checked.

In reality, methods start their life with purity of "unpredictable".
As execution progress, their purity are determined litte by little.
So a method that is already marked begin something other than
unpredictable need not be re-scaned, unless someone (re-)defined some
methods between the last scan and now.  To meet this situation we use
the global VM timestamp, and stores the last timestamp into an ISeq.