local variable usage investigation
The motivation behind this is that we want to eliminate local
variables if possible.  In order to do so the "possibility" of
variable elimination must be investigated beforehand, so this commit.

Any sort of investigations can be thought of but in this commit only
one (or two, technically speaking) kinds of characteristics are
collected against all local variables:

- If that local variable is read, and
- If that local variable is written.

It must be noted that in Ruby, there are no such thing like a read-
only variable.  All variables must be written before read.  So every
local variables are at least once written to exist.  "If the local
variable is written" must always be true.  What makes sense instead is
if the variable is read or not.  A write-only variable is possible,
and of course, subject to be eliminated.

This investigation is of course an overhead.  It slows down
compilation of a script so the main victim of this effect is of course
eval.  Slowing that method is not a bad idea anyway :trollface:, and
the good news is, the cost is very faint.

-----------------------------------------------------------
benchmark results:
minimum results in each 3 measurements.
Execution time (sec)
name    trunk   after
vm2_eval*       32.168  33.064

Speedup ratio: compare with the result of `trunk' (greater is better)
name    after
vm2_eval*       0.973