define purity of each instructions
This changeset introduces new instruction attribute called "purity".
By doing so we can eliminate calls to methods that are entirely
consist of pure instructions.

The definition of purity is chosen to achieve that optimisation; that
is, only instructions that do noting except stack manipulations are
marked so.  For instance instruction `once` is not pure, because it
can block other threads (so there can potentially be global side
effects).

A method call can both be pure and not pure at the same time.  What is
called at a specific call site cannot be determined until the very
moment when we actually call it.  Of course a method cannot be pure,
until every and all of methods it calls are (possibly recursively)
pure.  So in short, purity of a method is updated on-the-fly.