vm.inc now in C99
This changeset modifies the VM generator so that vm.inc is written
in C99.  Also added some comments in _insn_entry.erb so that the
intention of each parts to be made more clear.  I think this
improves overall readability of the generated VM.

Also confirmed that the exact same binary is generated before/after
this changeset.